hwf4 usart: switch from size_t types to int types
* this isnt kosher c but its required for ivory compatibility

+++ bsp/hwf4/include/hwf4/usart.h
@@ -40,20 +40,20 @@ void usart_enable(struct usart *usart);
 /**
  * Write to the usart, delaying for at most timeout*len ms.
  */
-ssize_t usart_write_timeout(struct usart *usart, uint32_t timeout,
-                            const uint8_t *buf, size_t len);
+int32_t usart_write_timeout(struct usart *usart, uint32_t timeout,
+                            const uint8_t *buf, uint32_t len);
 
 /**
  * Read from the usart, delaying for at most timeout*len ms.
  */
-ssize_t usart_read_timeout(struct usart *usart, uint32_t timeout, uint8_t *buf,
-                           size_t len);
+int32_t usart_read_timeout(struct usart *usart, uint32_t timeout, uint8_t *buf,
+                           uint32_t len);
 
 /**
  * Write to the usart device, blocking until the entire buffer has been written.
  */
-static inline ssize_t usart_write(struct usart *usart, const uint8_t *buf,
-                                  size_t len) {
+static inline int32_t usart_write(struct usart *usart, const uint8_t *buf,
+                                  uint32_t len) {
     return usart_write_timeout(usart, portMAX_DELAY, buf, len);
 }
 
@@ -62,21 +62,21 @@ static inline ssize_t usart_write(struct usart *usart, const uint8_t *buf,
  * the internal buffer to hold the data.  This is useful for
  * debugging, but should not be relied upon in normal operation.
  */
-void usart_write_from_isr(struct usart *usart, const uint8_t *buf, size_t len);
+void usart_write_from_isr(struct usart *usart, const uint8_t *buf, uint32_t len);
 
 /**
  * Perform a blocking write directly to the USART without interrupts.
  * Do not combine calls to this with normal writes.
  */
-ssize_t usart_write_blocking(struct usart *usart, const uint8_t *buf,
-                             size_t len);
+int32_t usart_write_blocking(struct usart *usart, const uint8_t *buf,
+                             uint32_t len);
 
 /**
  * Read from the usart device, blocking until the entire buffer has been
  * written.
  */
-static inline ssize_t usart_read(struct usart *usart, uint8_t *buf,
-                                 size_t len) {
+static inline int32_t usart_read(struct usart *usart, uint8_t *buf,
+                                 uint32_t len) {
     return usart_read_timeout(usart, portMAX_DELAY, buf, len);
 }
 
@@ -90,7 +90,7 @@ bool usart_is_tx_pending(struct usart *usart);
  * becomes available asynchronously via the RX interrupt.  It can be
  * used to poll for input.
  */
-size_t usart_available(struct usart *usart);
+uint32_t usart_available(struct usart *usart);
 
 /**
  * Read the first byte of input without removing it.
@@ -106,7 +106,7 @@ bool usart_peek(struct usart *usart, uint8_t *buf);
  * This number is useful as a lower bound on the value---it will
  * decrease asynchronously via the TX interrupt.
  */
-size_t usart_txspace(struct usart *usart);
+uint32_t usart_txspace(struct usart *usart);
 
 /* USART Handles **************************************************************/
